+++ CHANGELOG.md
@@ -63,10 +63,14 @@ Please see the [Envoy documentation](https://www.envoyproxy.io/docs/envoy/latest
 
 ## RELEASE NOTES
 
-## Next Release
+## [1.13.7] June 03, 2021
+[1.13.7]: https://github.com/datawire/ambassador/compare/v1.13.6...v1.13.7
+
+### Emissary Ingress and Ambassador Edge Stack
 
 - Feature: Add AMBASSADOR_JSON_LOGGING to enable JSON for most of the Ambassador control plane. Some (but few) logs from gunicorn and the Kubernetes client-go package still log text.
 - Bugfix: Fixed a bug where the Consul resolver would not actually use Consul endpoints with TCPMappings.
+- Change: Ambassador now calculates its own memory usage in a way that is more similar to how the kernel OOMKiller tracks memory.
 
 ## [1.13.6] May 24, 2021
 [1.13.6]: https://github.com/datawire/ambassador/compare/v1.13.5...v1.13.6